==> p11_plot_scatter.py <==
#!/usr/bin/env python
# -*- coding: utf-8 -*-

"""

This script visualises results generated by the script p09 as scatter plots.

"""

__author__ = "Manuel Linsenmeier"
__email__ = "m.linsenmeier@lse.ac.uk"
__version__ = "0.9"

import sys
import os
import copy
from copy import deepcopy

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import matplotlib as mpl
import seaborn as sns
from matplotlib.ticker import FormatStrFormatter

from scipy import stats

def r2(x, y):
	return stats.pearsonr(x, y)[0] ** 2

## ============================================================

dependent = 'length_sequence'
treatment = 'year_adoption'
df_residuals = pd.read_csv('./results/residuals_sequence_sample_all_year.csv')
df_residuals = df_residuals.set_index('iso')
countries = df_residuals.index.values

fig, ax = plt.subplots(figsize=(4,3))
for i, iso in enumerate(countries):
	x = df_residuals.loc[iso, treatment]
	y = df_residuals.loc[iso, dependent]
	ax.plot(x, y, color='white')
	ax.annotate(s=iso, xy=(x, y), xycoords='data', ha='center', va='center', fontsize='small')
xall = df_residuals[treatment].values
yall = df_residuals[dependent].values
sns.regplot(x=xall, y=yall, ax=ax, scatter=False, color='grey', order=1, ci=None, label='R2={0:3.2f}'.format(r2(xall, yall)))
ax.xaxis.set_major_formatter(FormatStrFormatter('%.0f'))
sns.despine(ax=ax, offset=1., right=True, top=True)
ax.legend()
ax.set_xlabel('Year of first adoption')
ax.set_ylabel('Length of policy sequence\n (partial residual)')
ax.set_xlim(2002, 2021)
fig.savefig('./figures/scatterplot_sequence.pdf', bbox_inches='tight')

## ============================================================

dependent = 'intensity'
treatment = 'length_sequence'
df_residuals = pd.read_csv('./results/residuals_pricelevel_sample_all_length.csv')
df_residuals = df_residuals.set_index('iso')

fig, ax = plt.subplots(figsize=(4,3))
for i, iso in enumerate(countries):
	x = df_residuals.loc[iso, treatment]
	y = df_residuals.loc[iso, dependent]
	ax.plot(x, y, color='white')
	ax.annotate(s=iso, xy=(x, y), xycoords='data', ha='center', va='center', fontsize='small')
xall = df_residuals[treatment].values
yall = df_residuals[dependent].values
sns.regplot(x=xall, y=yall, ax=ax, scatter=False, color='grey', order=1, ci=None, label='R2={0:3.2f}'.format(r2(xall, yall)))
ax.xaxis.set_major_formatter(FormatStrFormatter('%.1f'))
sns.despine(ax=ax, offset=1., right=True, top=True)
ax.legend()
ax.set_xlabel('Length of policy sequence')
ax.set_ylabel('Price intensity (partial residual)')
ax.set_xlim(1.5, 7.2)
sns.despine(ax=ax, offset=1., right=True, top=True)
fig.savefig('./figures/scatterplot_pricelevel.pdf', bbox_inches='tight')
